Item comments
Item folder done with comments

+++ TornApiProxy/Contract/Item/MarketListing.cs
@@ -20,19 +20,31 @@
 {
     using Newtonsoft.Json;
 
+    /// <summary>
+    /// A listing in a market
+    /// </summary>
     public class MarketListing
     {
+        /// <summary>
+        /// The cost of them item
+        /// </summary>
         [JsonProperty("cost")]
-        public int Cost { get; set; }
+        public long Cost { get; set; }
 
+        /// <summary>
+        /// The quantity of the item for sale
+        /// </summary>
         [JsonProperty("quantity")]
         public int Quantity { get; set; }
     }
 
+    /// <summary>
+    /// A subclass which is specific to points on the market
+    /// </summary>
     public class PointsListing : MarketListing
     {
         /// <summary>
-        /// Only applies to PointsMarket
+        /// Only applies to PointsMarket and represents the total cost to buy
         /// </summary>
         [JsonProperty("total_cost")]
         public int TotalCost { get; set; }
